requirements.txt
python-telegram-bot~=21.10
python-decouple~=3.8
pandas~=2.2.3
pip-review
gspread~=6.1.4
oauth2client~=4.1.3
wakeonlan~=3.1.0
psutil~=6.1.1